ipfixprobe - introduce process VLAN plugin
