+++ src/plugins/process/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(basicplus)
 add_subdirectory(bstats)
 add_subdirectory(icmp)
+add_subdirectory(vlan)
